Домашка Добавлен второй сканер для HDD
второй сканер удалён за ненадобностью

CPU + HDD + Monitor

методы печати вынесены в отдельный класс class A,
добавлена проверка типа диска, в случае если это SSD (ssd) предлагается ввести скорость вращения = 0

незначительные изменения

+++ src/day4/Monitor.java
@@ -10,7 +10,7 @@ public class Monitor {
     private int monResolutionY; //разрешение монитора по Y
     private String nameMonitor;  //название монитора
 
-    public double getMonResolutionX() {
+    public int getMonResolutionX() {
         return monResolutionX;
     }
